Rename regrid_method_type to dataclass_type and move _RegridVarType to regrid_schemes.py
