+++ imod/common/utilities/regrid.py
@@ -16,7 +16,7 @@ from imod.common.interfaces.iregridpackage import IRegridPackage
 from imod.common.interfaces.isimulation import ISimulation
 from imod.common.statusinfo import NestedStatusInfo
 from imod.common.utilities.clip import clip_by_grid
-from imod.common.utilities.regrid_method_type import EmptyRegridMethod, RegridMethodType
+from imod.common.utilities.dataclass_type import EmptyRegridMethod, RegridMethodType
 from imod.common.utilities.value_filters import is_valid
 from imod.mf6.validation_settings import ValidationSettings
 from imod.schemata import ValidationError
