+++ imod/mf6/dis.py
@@ -12,7 +12,7 @@ from imod.common.utilities.regrid import (
     _regrid_like,
     _regrid_package_data,
 )
-from imod.common.utilities.regrid_method_type import RegridMethodType
+from imod.common.utilities.dataclass_type import RegridMethodType
 from imod.logging import init_log_decorator, standard_log_decorator
 from imod.mf6.disv import VerticesDiscretization
 from imod.mf6.package import Package
