+++ imod/mf6/model_gwf.py
@@ -7,7 +7,7 @@ from typing import Optional, cast
 import cftime
 import numpy as np
 
-from imod.common.utilities.regrid_method_type import RegridMethodType
+from imod.common.utilities.dataclass_type import RegridMethodType
 from imod.logging import init_log_decorator
 from imod.logging.logging_decorators import standard_log_decorator
 from imod.mf6 import ConstantHead
